requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
av
certifi
cffi
charset-normalizer
click
cryptography
colorama
coloredlogs
distro
docstring_parser
eval_type_backport
filelock
flatbuffers
frozenlist
fsspec
h11
httpcore
httpx
huggingface-hub
humanfriendly
idna
Jinja2
jiter
livekit
livekit-agents
livekit-api
livekit-plugins-cartesia
livekit-plugins-noise-cancellation
livekit-plugins-openai
livekit-plugins-rime
livekit-plugins-silero
livekit-plugins-tavus
livekit-plugins-turn-detector
livekit-protocol
MarkupSafe
mpmath
multidict
nest-asyncio
numpy
onnxruntime
openai
packaging
pillow
propcache
protobuf
psutil
pycparser
pydantic
pydantic_core
PyJWT
python-dotenv
PyYAML
regex
requests
safetensors
sniffio
snowflake-connector-python
sounddevice
sympy
tokenizers
torch
tqdm
transformers
types-protobuf
typing-inspection
typing_extensions
Unidecode
urllib3
watchfiles
websockets
yarl
